--- graphql/.eslintrc
@@ -1,5 +1,0 @@
-{
-  "rules": {
-    "import/no-extraneous-dependencies": "OFF"
-  }
-}
