+++ src/components/OpenSourceContributions/OpenSourceContributionsItem.js
@@ -1,15 +1,15 @@
 import React from 'react';
-import Text from 'components/Text';
+import GithubRepository from 'components/GithubRepository';
 import Icon from 'components/Icon';
 import Link from 'components/Link';
-import GithubRepository from 'components/GithubRepository';
+import Text from 'components/Text';
 import styles from './OpenSourceContributionsItem.module.scss';
 
 export default function OpenSourceContributionsItem({contribution}) {
   return (
     <GithubRepository repository={contribution.repository}>
       <div className={styles.pullRequests}>
-        {contribution.pullRequests.map(pullRequest => (
+        {contribution.pullRequests.map((pullRequest) => (
           <Text key={pullRequest.id} className={styles.pullRequest}>
             <Icon
               className={styles.pullRequestIcon}
