+++ src/config.js
@@ -1,0 +1,3 @@
+export default {
+  GH_TOKEN: process.env.GH_TOKEN
+};
